jdbctemplate

+++ src/main/java/hello/hellospring/repository/JdbcMemberRepository.java
@@ -32,7 +32,7 @@ public class JdbcMemberRepository implements MemberRepository{
             pstmt.executeUpdate();
             rs = pstmt.getGeneratedKeys();
             if (rs.next()) {
-                member.setId(rs.getInt(1));
+                member.setId(rs.getLong(1));
             } else {
                 throw new SQLException("id 조회 실패");
             }
@@ -45,7 +45,7 @@ public class JdbcMemberRepository implements MemberRepository{
     }
 
     @Override
-    public Optional<Member> findById(int id) {
+    public Optional<Member> findById(Long id) {
         String sql = "select * from member where id = ?";
         Connection conn = null;
         PreparedStatement pstmt = null;
@@ -53,11 +53,11 @@ public class JdbcMemberRepository implements MemberRepository{
         try {
             conn = getConnection();
             pstmt = conn.prepareStatement(sql);
-            pstmt.setInt(1, id);
+            pstmt.setLong(1, id);
             rs = pstmt.executeQuery();
             if(rs.next()) {
                 Member member = new Member();
-                member.setId(rs.getInt("id"));
+                member.setId(rs.getLong("id"));
                 member.setName(rs.getString("name"));
                 return Optional.of(member);
             } else {
@@ -82,7 +82,7 @@ public class JdbcMemberRepository implements MemberRepository{
             pstmt.setString(1, name); rs = pstmt.executeQuery();
             if(rs.next()) {
                 Member member = new Member();
-                member.setId(rs.getInt("id"));
+                member.setId(rs.getLong("id"));
                 member.setName(rs.getString("name"));
                 return Optional.of(member);
             }
@@ -106,7 +106,7 @@ public class JdbcMemberRepository implements MemberRepository{
             List<Member> members = new ArrayList<>();
             while(rs.next()) {
                 Member member = new Member();
-                member.setId(rs.getInt("id"));
+                member.setId(rs.getLong("id"));
                 member.setName(rs.getString("name"));
                 members.add(member);
             }
