+++ src/main/java/hello/hellospring/repository/MemoryMemberRepository.java
@@ -7,8 +7,8 @@ import java.util.*;
 
 public class MemoryMemberRepository implements MemberRepository{
 
-    private static Map<Integer, Member> store = new HashMap<>();
-    private static int sequence = 0;
+    private static Map<Long, Member> store = new HashMap<>();
+    private static Long sequence = 0L;
 
     @Override
     public Member save(Member member) {
@@ -18,7 +18,7 @@ public class MemoryMemberRepository implements MemberRepository{
     }
 
     @Override
-    public Optional<Member> findById(int id) {
+    public Optional<Member> findById(Long id) {
 
         return Optional.ofNullable(store.get(id));
     }
